Integrating out-of-core mechanism with credit-based flow-control and data generation tethering
Summary: This diff integrates out-of-core infrastructure with credit-based flow control and adds the ability to tether the rate of data generation/processing. Data generation/processing rate is controlled by changing the number of active processing (input/compute) threads. This diff also implements a new (and more performant) adaptive out-of-core policy.

Test Plan:
mvn clean verify
all snapshot tests including ones with large data pass
Running adaptive out-of-core on large graph with very limited memory does not fail.
This diff should enable us to avoid *any* reasonable job to fail!

Reviewers: maja.kabiljo, sergey.edunov, avery.ching, dionysis.logothetis

Reviewed By: dionysis.logothetis

Subscribers: ramesh-muthusamy

Differential Revision: https://reviews.facebook.net/D55479

+++ giraph-core/src/main/java/org/apache/giraph/comm/netty/handler/ResponseClientHandler.java
@@ -64,11 +64,11 @@ public class ResponseClientHandler extends ChannelInboundHandlerAdapter {
     ByteBuf buf = (ByteBuf) msg;
     int senderId = -1;
     long requestId = -1;
-    short response = -1;
+    int response = -1;
     try {
       senderId = buf.readInt();
       requestId = buf.readLong();
-      response = buf.readShort();
+      response = buf.readInt();
     } catch (IndexOutOfBoundsException e) {
       throw new IllegalStateException(
           "channelRead: Got IndexOutOfBoundsException ", e);
